+++ resources/views/admin/reviews/index.blade.php
@@ -5,7 +5,7 @@
             <tr>
                 <th class="px-4 py-2">ID</th>
                 <th class="px-4 py-2">User</th>
-                <th class="px-4 py-2">Object</th>
+                <th class="px-4 py-2">Category</th>
                 <th class="px-4 py-2">Rating</th>
                 <th class="px-4 py-2">Actions</th>
             </tr>
@@ -15,7 +15,7 @@
             <tr class="border-t">
                 <td class="px-4 py-2">{{ $review->id }}</td>
                 <td class="px-4 py-2">{{ $review->user->name }}</td>
-                <td class="px-4 py-2">{{ $review->object->title }}</td>
+                <td class="px-4 py-2">{{ $review->category->title }}</td>
                 <td class="px-4 py-2">{{ $review->rating }}</td>
                 <td class="px-4 py-2 space-x-2">
                     <a href="{{ route('admin.reviews.edit', $review) }}" class="text-indigo-600">Edit</a>
