+++ resources/views/profile/show.blade.php
@@ -45,7 +45,6 @@
         <div class="bg-white shadow rounded-lg">
             <nav class="border-b flex space-x-8 px-6 pt-4">
                 <a href="#tab-about" class="tab-link text-gray-600 hover:text-indigo-600 font-medium" data-tab="about">Обо мне</a>
-                <a href="#tab-objects" class="tab-link text-gray-600 hover:text-indigo-600 font-medium" data-tab="objects">Объекты ({{ $objects->count() }})</a>
                 <a href="#tab-reviews" class="tab-link text-gray-600 hover:text-indigo-600 font-medium" data-tab="reviews">Отзывы ({{ $reviews->count() }})</a>
                 <a href="#tab-comments" class="tab-link text-gray-600 hover:text-indigo-600 font-medium" data-tab="comments">Комментарии ({{ $comments->count() }})</a>
                 @auth
@@ -62,27 +61,12 @@
                 <div id="tab-about" class="tab-content">
                     <p class="text-gray-700">{{ $user->about ?? 'Пользователь ещё не рассказал о себе.' }}</p>
                 </div>
-                <div id="tab-objects" class="tab-content hidden space-y-4">
-                    @forelse($objects as $object)
-                        <div class="bg-gray-50 p-4 rounded-md border">
-                            <div class="flex items-center justify-between">
-                                <a href="{{ route('objects.show', $object->slug) }}" class="text-indigo-600 hover:underline font-medium">
-                                    {{ $object->title }}
-                                </a>
-                                <span class="text-gray-500 text-sm">{{ $object->created_at->format('d.m.Y') }}</span>
-                            </div>
-                            <p class="mt-1 text-gray-700">Категория: {{ $object->category?->title }}</p>
-                        </div>
-                    @empty
-                        <p class="text-gray-600 italic">Объектов пока нет.</p>
-                    @endforelse
-                </div>
                 <div id="tab-reviews" class="tab-content hidden space-y-4">
                     @forelse($reviews as $review)
                         <div class="bg-gray-50 p-4 rounded-md border">
                             <div class="flex items-center justify-between">
-                                <a href="{{ route('objects.show', $review->object->slug) }}" class="text-indigo-600 hover:underline font-medium">
-                                    {{ $review->object->title }}
+                                <a href="{{ route('categories.show', $review->category->slug) }}" class="text-indigo-600 hover:underline font-medium">
+                                    {{ $review->category->title }}
                                 </a>
                                 <span class="text-gray-500 text-sm">{{ $review->created_at->format('d.m.Y') }}</span>
                             </div>
@@ -100,7 +84,7 @@
                     @forelse($comments as $comment)
                         <div class="bg-gray-50 p-4 rounded-md border">
                             <div class="flex items-center justify-between">
-                                <a href="{{ route('objects.show', $comment->review->object->slug) }}#review-{{ $comment->review->id }}" class="text-indigo-600 hover:underline font-medium">
+                                <a href="{{ route('categories.show', $comment->review->category->slug) }}#review-{{ $comment->review->id }}" class="text-indigo-600 hover:underline font-medium">
                                     К отзыву: {{ \Illuminate\Support\Str::limit($comment->review->content, 50) }}
                                 </a>
                                 <span class="text-gray-500 text-sm">{{ $comment->created_at->format('d.m.Y H:i') }}</span>
